add content hash to URLs in css
when a url is referenced in css:

    background-image: url(/static/img/logo.png);

there was no way to communicate to the browser when the image had
changed. so if we were to update the storediq logo, for instance, since
the URL didn't change and we've got the cache expiration set to one year
in nginx, browsers wouldn't fetch the new logo.

this problem doesn't crop up very often with stuff like the app logo,
but it's a much more significant problem in the case of the icon fonts,
which will be updated relatively frequently with new icons.

the solution is to add a content hash to the url:

    background-image: url(/static/img/logo.png?contenthash=a67b58);

the content hash is re-calculated every build, so browsers will only
re-fetch an image or font if it has actually been updated.

pertains to DAQ-447